Disable blur on PopupButton when modal is open
