+++ package.json
@@ -165,7 +165,7 @@
         "@togglecorp/fujs": "^1.9.2-beta.0",
         "@togglecorp/re-map": "^0.1.3",
         "@togglecorp/toggle-form": "^0.2.0-beta.6",
-        "@togglecorp/toggle-ui": "^0.16.0-beta.0",
+        "@togglecorp/toggle-ui": "^0.16.0-beta.1",
         "@turf/bbox": "^6.0.1",
         "@turf/bbox-polygon": "^6.0.1",
         "@turf/combine": "^6.0.1",
